requirements.txt
flask==2.0.1
pillow==9.0.1
torch==1.10.0
torchvision==0.11.1